upgrade nextjs and deps
