--- apps/kiosk-web/app/controller/app/head.tsx
@@ -1,7 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>แผงควบคุม PM Station Kiosk</title>
-    </>
-  );
-}
