--- apps/kiosk-web/app/controller/app/page.tsx
@@ -1,6 +1,0 @@
-import { redirect } from "next/navigation";
-
-export default function ControllerAppPage() {
-  redirect("/controller/app/queue");
-  return null;
-}
