--- apps/kiosk-web/app/controller/app/queue/head.tsx
@@ -1,1 +1,0 @@
-export { default } from "../head";
