+++ apps/kiosk-web/app/controller/app/route.ts
@@ -1,0 +1,7 @@
+import { NextRequest, NextResponse } from "next/server";
+
+export const GET = (request: NextRequest) => {
+  const url = request.nextUrl.clone();
+  url.pathname = "/controller/app/queue";
+  return NextResponse.redirect(url);
+};
