--- apps/kiosk-web/app/controller/head.tsx
@@ -1,10 +1,0 @@
-export default function Head() {
-    return (
-      <>
-        <title>
-          Login : PM Station Kiosk
-        </title>
-      </>
-    );
-  }
-
