+++ apps/kiosk-web/app/layout.tsx
@@ -1,4 +1,5 @@
 import { Prompt } from "@next/font/google";
+import { Metadata } from "next";
 import "react-toastify/dist/ReactToastify.min.css";
 import "../styles/globals.css";
 import "../styles/pm-station.css";
@@ -9,6 +10,13 @@ const promptFont = Prompt({
   subsets: ["latin", "thai"],
   display: "block",
 });
+
+export const metadata: Metadata = {
+  title: {
+    default: "PM Station Kiosk",
+    template: "%s | PM Station Kiosk",
+  },
+};
 
 export default function RootLayout({
   children,
