--- apps/kiosk-web/app/projector/head.tsx
@@ -1,9 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>
-        โครงการ PM Station โดยคณะกรรมการนักเรียนประจำปีการศึกษา 2565
-      </title>
-    </>
-  );
-}
