--- apps/kiosk-web/app/้head.tsx
@@ -1,10 +1,0 @@
-export default function Head() {
-  return (
-    <>
-      <title>
-        โครงการ PM Station โดยคณะกรรมการนักเรียนประจำปีการศึกษา 2565
-      </title>
-      <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    </>
-  );
-}
