+++ apps/kiosk-web/package.json
@@ -13,7 +13,7 @@
   "dependencies": {
     "@googleapis/drive": "4.0.2",
     "@headlessui/react": "1.7.10",
-    "@heroicons/react": "^2.0.13",
+    "@heroicons/react": "^2.0.17",
     "@next/font": "13.1.6",
     "@station/client": "*",
     "@station/server": "*",
@@ -23,9 +23,11 @@
     "framer-motion": "^8.2.4",
     "kiosk-socket": "*",
     "logrocket": "^3.0.1",
-    "next": "13.1.6",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "next": "13.4.2",
+    "pusher": "^5.1.3",
+    "pusher-js": "^8.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-modal-sheet": "^1.9.0",
     "react-toastify": "9.1.1",
     "react-use-audio-player": "^1.2.6",
@@ -37,13 +39,13 @@
   },
   "devDependencies": {
     "@types/howler": "^2.2.7",
-    "@types/node": "18.11.17",
-    "@types/react": "18.0.26",
-    "@types/react-dom": "18.0.10",
-    "autoprefixer": "^10.4.13",
+    "@types/node": "18.16.3",
+    "@types/react": "18.2.0",
+    "@types/react-dom": "18.2.1",
+    "autoprefixer": "10.4.14",
     "build-config": "*",
-    "postcss": "^8.4.18",
-    "tailwindcss": "^3.2.4",
-    "typescript": "4.9.5"
+    "postcss": "8.4.23",
+    "tailwindcss": "3.3.2",
+    "typescript": "5.0.4"
   }
 }
